server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
slowapi
python-dotenv
pydantic-settings
pyjwt
passlib
bcrypt==3.2.2
python-multipart
email-validator
httpx
pytest